perf(tests): split the ce-work recovery shard, the largest remaining one
Address PR review feedback (#1370).

The first split landed unevenly: recovery was 35.6s of CPU against 22.7s,
18.7s, and 10.5s for its siblings, which made it a co-critical-path file
alongside ce-babysit-pr-snapshot. Split it into fallback (15 tests) and
retries (10 tests). The five shards now run 27.3s wall together, down
from 38.8s.

Also replaces the bare thousand-line rule this PR added with the
criterion it was standing in for. Line count is a smell; measured
per-file time against the suite's slowest-file ceiling is the actual
test. transport (1,071 lines, 18.7s) and verification (1,219 lines,
22.7s) sit well under the ~36s ceiling, so they stay whole.
